feat: zero_adder function
