+++ src/menu/binary/helper/helper.h
@@ -1,0 +1,13 @@
+#ifndef HELPER_H
+#define HELPER_H
+#include "is_negative_binary.c"
+#include "is_negative_binary_with_dot.c"
+#include "is_positive_binary.c"
+#include "is_positive_binary_with_dot.c"
+
+extern int is_negative_binary(char *binary_input);
+extern int is_negative_binary_with_dot(char *binary_input);
+extern int is_positive_binary(char *binary_input);
+extern int is_positive_binary_with_dot(char *binary_input);
+
+#endif
